requirements.txt
python-dotenv
ipykernel
ipywidgets
pandas
datasets
inspect-ai
openai
anthropic
google-api-python-client
tavily-python
pyyaml